14.6 hero animation
